api# inline versioning - review comments
