+++ api/src/main/java/marquez/db/DatasetVersionDao.java
@@ -23,11 +23,8 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.NonNull;
 import marquez.common.Utils;
-import marquez.common.models.DatasetName;
-import marquez.common.VersionUtils;
 import marquez.common.models.Field;
 import marquez.common.models.FieldName;
-import marquez.common.models.NamespaceName;
 import marquez.common.models.RunId;
 import marquez.common.models.TagName;
 import marquez.common.models.Version;
@@ -67,7 +64,7 @@ public interface DatasetVersionDao extends BaseDao {
     TagDao tagDao = createTagDao();
     DatasetFieldDao datasetFieldDao = createDatasetFieldDao();
 
-    Version version = VersionUtils.newDatasetVersionFor(namespaceName, datasetName, datasetMeta);
+    Version version = Utils.newDatasetVersionFor(namespaceName, datasetName, datasetMeta);
     UUID newDatasetVersionUuid = UUID.randomUUID();
     DatasetVersionRow datasetVersionRow =
         upsert(
