+++ integrations/spark/src/main/java/marquez/spark/agent/lifecycle/plan/SaveIntoDataSourceCommandVisitor.java
@@ -61,7 +61,8 @@ public class SaveIntoDataSourceCommandVisitor
                 relationVisitors,
                 new LogicalRelation(
                     relation, relation.schema().toAttributes(), Option.empty(), x.isStreaming())))
-        .orElse(Collections.emptyList()).stream()
+        .orElse(Collections.emptyList())
+        .stream()
         // constructed datasets don't include the output stats, so add that facet here
         .peek(
             ds -> {
